+++ Source/WebGPU/WGSL/AST/ASTStatement.h
@@ -32,50 +32,36 @@
 
 namespace WGSL::AST {
 
-class Expression : public ASTNode {
+class Statement : public ASTNode {
     WTF_MAKE_FAST_ALLOCATED;
 
 public:
-    enum class Kind {
-        BoolLiteral,
-        Int32Literal,
-        Uint32Literal,
-        Float32Literal,
-        AbstractIntLiteral,
-        AbstractFloatLiteral,
-        Identifier,
-        ArrayAccess,
-        StructureAccess,
-        CallableExpression,
-        UnaryExpression,
+    enum class Kind : uint8_t {
+        Compound,
+        Return,
+        Assignment,
+        Variable,
     };
 
-    using List = UniqueRefVector<Expression>;
+    using List = UniqueRefVector<Statement>;
 
-    Expression(SourceSpan span)
+    Statement(SourceSpan span)
         : ASTNode(span)
     {
     }
 
-    virtual ~Expression() {}
+    virtual ~Statement() { }
 
     virtual Kind kind() const = 0;
-    bool isBoolLiteral() const { return kind() == Kind::BoolLiteral; }
-    bool isInt32Literal() const { return kind() == Kind::Int32Literal; }
-    bool isUInt32Literal() const { return kind() == Kind::Uint32Literal; }
-    bool isFloat32Literal() const { return kind() == Kind::Float32Literal; }
-    bool isAbstractIntLiteral() const { return kind() == Kind::AbstractIntLiteral; }
-    bool isAbstractFloatLiteral() const { return kind() == Kind::AbstractFloatLiteral; }
-    bool isIdentifier() const { return kind() == Kind::Identifier; }
-    bool isArrayAccess() const { return kind() == Kind::ArrayAccess; }
-    bool isStructureAccess() const { return kind() == Kind::StructureAccess; }
-    bool isCallableExpression() const { return kind() == Kind::CallableExpression; }
-    bool isUnaryExpression() const { return kind() == Kind::UnaryExpression; }
+    bool isCompound() const { return kind() == Kind::Compound; }
+    bool isReturn() const { return kind() == Kind::Return; }
+    bool isAssignment() const { return kind() == Kind::Assignment; }
+    bool isVariable() const { return kind() == Kind::Variable; }
 };
 
 } // namespace WGSL::AST
 
-#define SPECIALIZE_TYPE_TRAITS_WGSL_EXPRESSION(ToValueTypeName, predicate) \
+#define SPECIALIZE_TYPE_TRAITS_WGSL_STATEMENT(ToValueTypeName, predicate) \
 SPECIALIZE_TYPE_TRAITS_BEGIN(WGSL::AST::ToValueTypeName) \
-    static bool isType(const WGSL::AST::Expression& expression) { return expression.predicate; } \
+    static bool isType(const WGSL::AST::Statement& statement) { return statement.predicate; } \
 SPECIALIZE_TYPE_TRAITS_END()
